Added endpoint security

+++ src/app.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI, File, UploadFile, Form, Depends, HTTPException
 from src.schemas import PostResponse, UserRead, UserCreate, UserUpdate
-from src.db import Post, create_db_and_tables, get_async_session
+from src.db import Post, create_db_and_tables, get_async_session, User
 from sqlalchemy.ext.asyncio import AsyncSession
 from contextlib import asynccontextmanager
 from sqlalchemy import select
@@ -30,6 +30,7 @@ app.include_router(fastapi_user.get_users_router(UserRead, UserUpdate), prefix='
 async def upload_post(
         file: UploadFile = File(...),
         caption: str = Form(''),
+        user: User = Depends(current_active_user),
         session: AsyncSession = Depends(get_async_session)
         ) -> PostResponse:
     temp_file_path = None
@@ -47,6 +48,7 @@ async def upload_post(
         )
         if upload_result.response_metadata.http_status_code == 200:
             post = Post(
+                user_id=user.id,
                 caption=caption,
                 url=upload_result.url,
                 file_type='video' if file.content_type.startswith('video') else 'image',
@@ -65,30 +67,42 @@ async def upload_post(
 
 @app.get('/feed')
 async def get_feed(
+        user: User = Depends(current_active_user),
         session: AsyncSession = Depends(get_async_session)
         ) -> dict:
     result = await session.execute(select(Post).order_by(Post.created_at.desc()))
-    posts = result.scalars().all()
+    posts = [row[0] for row in result.all()]
+    result = await session.execute(select(User))
+    users = [row[0] for row in result.all()]
+    user_dict = {u.id: u.email for u in users}
     post_data = [
         {
             'id': str(post.id),
+            'user_id': str(post.user_id),
             'caption': post.caption,
             'url': post.url,
             'file_type': post.file_type,
             'file_name': post.file_name,
-            'created_at': post.created_at.isoformat()
+            'created_at': post.created_at.isoformat(),
+            'is_owner': post.user_id == user.id,
+            "email": user_dict.get(post.user_id, "Unknown")
         }
         for post in posts
     ]
     return {'posts': post_data}
 
 @app.delete('/post/{post_id}')
-async def delete_post(post_id: str, session: AsyncSession = Depends(get_async_session)) -> dict[str, bool]:
+async def delete_post(
+        post_id: str,
+        user: User = Depends(current_active_user),
+        session: AsyncSession = Depends(get_async_session)
+    ) -> dict[str, bool]:
     try:
         post_uuid = UUID(post_id)
         result = await session.execute(select(Post).where(Post.id == post_uuid))
         post = result.scalars().first()
         if not post: raise HTTPException(status_code=404, detail='Post ID wrong!')
+        if post.user_id != user.id: raise HTTPException(status_code=403, detail="You don't have permission!")
         await session.delete(post)
         await session.commit()
         return {'success': True}
